Merge branches 'Authentication' and 'master' of https://github.com/mbeck5/GU_Energy_Dashboard into Authentication
Conflicts:
	client/app/index.html

Converted login to be a modal.  The login button is now in the footer of the competition list.  When the user is not logged in, the login button shows there.  When the user is logged in, the create/edit/delete buttons are shown there.  Also, the logout button appears in the header when the user is logged in.  Clicking that log the user out and change the competition footer buttons accordingly.